backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
supabase>=2.25.1
PyJWT>=2.10.1
python-dotenv>=1.0.1
pydantic>=2.10.3
httpx>=0.28.1
python-multipart>=0.0.20
